+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Updates
 * Support Suspend and Resume Client APIs ([#104](https://github.com/microsoft/durabletask-java/issues/104))
+* Fix the potential NPE issue of `DurableTaskClient terminate` method ([#104](https://github.com/microsoft/durabletask-java/issues/104))
 
 
 ## v1.0.0